Import evolution requirements queries into main db file.
